requirements.txt
flask
python-dotenv
pandas
gtts
pydub
twilio
apscheduler
